Changed name of script / Finished code
I wanted to change the name of the file so I copy and pasted my previous code into a new document with a new name and deleted the old.

I made 2 different scripts because one of them runs when you just straight off press the arrow button but the other should run straight off "python lab2script.py"

As well as the final product.